Add --clean/--no-cleanup cli option
Former-commit-id: 3b1b2f4c5dcd37b75017526f8e584a6352c018b3

+++ src/job.ts
@@ -436,6 +436,8 @@ export class Job {
     async cleanupResources () {
         clearTimeout(this._longRunningSilentTimeout);
 
+        if (!this.argv.cleanup) return;
+
         for (const id of this._containersToClean) {
             try {
                 await Utils.spawn(["docker", "rm", "-f", `${id}`]);
